Use the rjil::system class
Contents for base user management had been
added to the rjil::system class, but it wasn't
actually being used.

This commit ensures that rjil::system is included
by default for all machines.

It also moved code to ensure that root logins are
disabled from the server class to the accounts class
and deleted the server class b/c it's locic had been
rewritten in accounts.