Move towards an eventually consistent cross host deployment model
The original design was based on the idea that all cross host
dependencies could be made in such a way such that they always
flowed in a single direction (ie: no circular dependencies).

After more considerations, this model is not going to support
some of our deployment requirements for openstack.

This patch makes 2 additional that convert the deployment process
from something strictly consistent to something eventually consistent.

1. make custom hiera backend way more forgiving

Previously, the backend would raise an exception and halt Puppet as
a catalog compilation failure. This meant that each host had to be
able to configure itself in one single run. Now, failure to look
up array data just means that it reverts to whatever defaults it
may need. This means that Puppet can perform partial runs, ensuring
that the service discovery file gets written even when the required
data is not ready.

2. publish services to etcd immeditately after run

so that even a run missing data (or one doomed to fail b/c of missing
requirements) can still publish its service for other services to
avoid deadlock.

NOTE: this solution will work, but it is not quite perfect. My main
concern with this approach is that it requires puppet runs to fail
multiple times before eventually resulting in a successful run. In essence,
the runs are eventually consistent (meaning that they will fail on each
node until enough data is ready for the cluster to activate). My concern
with this model is that not all failure states are going to be recoverable,
and it will be really hard to understand expected vs. unexpected failure
states. This is the exact reason that we are looking at moving all cross
host coordination to etcd. My other concern is that the results are
possibly non-deterministic. Depending on the order in which things happen
across multiple hosts, the same code could result in either a functional
or a non-functional cluster. ALthough I don't have evidence that this is
true, I also don't feel that it can be guarenteed these things won't
happen.