Add jiocloud_role fact to hierarchy
This patch adds a new fact called jiocloud_role.

This fact just parses the mathing role out of the
hostname.

It adds this role to the hierarchy as role-env
which takes precedence over env. This means that
you can add role specfic data that applies
only to a specific environment.

This hierarchy was added for cases where a service
might need to resolve data from itself as opposed
to from the service discovery data it is happans
to be providing services itself.