Allow PR to be set for repo
Using the github pull request builder, it is
not aware of the remote repo and branch for
a pull request, it only makes the pull request
available.

This patch allows the code that to use for puppet-rjil
for each node to be determined by the pull request id
and upstream repo name.