Support ceph storage cluster less than 3 nodes
By default ceph cluster need at least three nodes to be in the
cluster to store data in it, as default number of object replicas
is 3.
Adding a parameter to change number of object replicas of all pools,
so that anyone need less than 3 nodes in cluster can specify appropriate
replication factor (1 or 2) to make the cluster up.