Added domain search for resolver
Configured domains to be searched (node.consul and service.consul) in
resolvconf. This is required for matchmaker as matchmaker need the configure
the hostnames for all node types (like compute nodes, scheduler etc) which
should be same as the result of hostname -s on the server and this name must be
resolvable as the components will connect with the hostname configured in
matchmaker config.

This patch will allow resolving the names without providing domain names.