Fixed wrong condition to fail on journal size
Fail condition on osd journal size is kept just to make sure that there will be
enough disk space for ceph data. This is only going to happen on test systems
when autogenerate is true and people keep autodisk_size small.

It dont have anything to do with processorcount and because of processor count
in the codition, it is failing when I set autogenerate true on baremetal system
where processor count is large in number.