Disable GRO
The generic-receive-offload code paths in be2net causes kernel
panics.

Put in place an upstart job that runs for every interface and disables
GRO and disable it from Puppet as well on the first run.

It seemed that we needed to do this for all interfaces to actually fix
the problem. I suspect it has to do with the fact that eth2 and eth3 are
actually ports on the same NIC, so we weren't properly avoiding the GRO
code paths in the driver if it was left turned on for the other port.