Add cloud_provider
Currently, we conflate the stage of the
test (ie: at) with the provider where the
test will be run (like hp cloud vs. our
staging environmnet)

this patch aims to resolve this by differenting
between those things. The aim is to allow us to
seemlessly fall back to hp cloud if we have
issues with another enviornment