Added tempest configuration code
This patch will setup the resources to ::tempest and ::tempest::provision code
to work and then will call them with appropriate hiera data.

Note:

Glance module version has changed: Glance_image type in the stable/icehouse
version was not working because of wrong arguments passed to glance client. So
used later version which fixed this problem.

Added global scope for public_network_name in hiera as it is getting used in
multiple places.