Only use udlc for ntp server(on etcd node)
Using udlc on all servers were causing the server to sync with local clock
if upstream server is not available. This may be required for ntp server (on
acceptance test environment) to avoid waiting long time to finish ntp server
initialization (it may take even couple of hours in general to completely sync
an ntp server with its upstream ntp server until which it will not be able to provide
time).
It is not required for ntp clients. This patch will do so.

This should fix the delay on ceph validation checks because of clock skew.