remove purging from partial catalog updated
This patch removes purge from file resources
that were causing states to flap for runs that
perform partial catalogs applied.

The default for /etc/consul and /etc/apt/sources.list.d
was to purge unknown files. For our partial puppet runs,
this was causing it to remove files that needed to
be there, but that were not included as a part of the
rjil::system::apt or rjil::jiocloud classes (which
included the contrail source, as well as most consul
service checks)

This patch adds a File colection override to ensure that those
resources do not purge for these partial catalog applies.