Create private interface fact for compute hosts
A single fact cannot be trusted to get the private
address of compute hosts b/c it changes from some
interface to vhost0.

This patch adds a fact that will use vhost0 if it is
set and fall back to eth0,1 as the primary address
so that we can reliably figure out the private address
for compute hosts.