Changed the endpoints to use jiocloud.com domain
SSL need the urls to have the domain names match to the domain name in the
certificate. The easiest option here (which works on all environments including
production) is to change the endpoints to the names in jiocloud.com domain and
make appropriate cnames which points lb.<servicename>.service.consul to
appropritate <servicename>.jiocloud.com.

In this case, all we need to do is to create self signed certificate for test
systems and CA signed certificate for prod/staging without any other changes.

So this patch making below changes.

* Add cnames in dnsmasq to all endpoints to <service name>.jiocloud.com.
  * I have taken the names we used in previous implementation (like compute for
nova, volume for cinder etc.)
* Change the endpoint to the new names.
* Used the same name on all internal/public/admin urls
  * All the requests including internal requests should use the urls from load
    balancer which was not true because we were using
    real.<servicename>.service.consul for internal and admin endpoints
  * The public access to the APIs will be through NATed IPs (to public IPs)
    in which case even though it use same names, internal servers will be
    resolving it to internal names (using dnsmasq cnames) and external clients will
    be using public IPs from public dns servers
* Other option is to use a different name for internal/admin urls (in jiocloud.com domain iteself
  something like compute-internal.jiocloud.com).
* Changed dns/entry.pp to use dns_resolve from orchestration_utils instead of
  dns_a because dns_a was failing the execution if the address is not resolvable
  which was causing the entire deployment stuck there.