Remove old test file
Since the files /usr/lib/jiocloud/tests/neutron.sh are being used
to check services for consul, the file neutron-service was converted
to this name so that the file used for validation could be shared by
consul for it's sevice health check.

While the name of the file was changed, however, the old version
was not deleting, meaning that it will be left on disk until
servers are reprovisioned.

In general, we should clean up after ourselves, so that we can
upgrade without accumulating too much configuration drift (
in this case, where newly provisioned nodes do not match
upgraded nodes.)