Fix Puppet Deprecation Warnings
Currently, every Puppet run throws two warnings saying that Puppet is
using `modulepath` and `manifestdir` which are deprecated. This is
primarily because Puppetlabs is moving towards the use of Directory
Environments.

This patch disables the default per environment manifests and sets up
globals for modules and manifests pointing to their corresponding
overrides directory.