Added options for allocation pools to allocate subset of IPs to floating IPs
There would be situations when set of IPs in the floating network provided need to be set
aside for some purpose, so that they can be statically managed ( may be
something like certain number of IPs to be added to physical network devices or
something which cannot be managed with neutron.

This patch to support providing ip address range between which to allocate for
floating pool.