Add simple service feature
In some cases, we need to delay starting
an application during bootstrapping so that
we can perform configuration to prepare a cluster
before we can start the service.

In some extreme cases, have package installs start
services can lead to split-brain cases and cause issues
with data replicate and initialization acorss clusters.

This patch implements the ability to delay starting a
specific list of services (initially: zookeeper, cassandra,
contrail-schema, and contrail-api) and ensure that the package
will not attempt to start them until after the service resourcs
in puppet has run.