remove ceph specific repo
Currently, we are installing ceph version:
0.80.9. This version can currently be found
both in our custom provided ceph repo as well
as the ubuntu trust repo. Each of these repos
has a slightly differnt named version of these
packages which is causing spurious build failures.

This patch removes the custom ceph repo so that
we can just rely on the trusty repo for our ceph
installs.