Refactor neutron network creation test
Currently, it's possible for neutron to get stuck into
a state where it continuously creates test networks
forever. This is a symptom of a fundamental probleam of
the networking script, that it does not keep track of the
exact objects being created to ensure only those exact objects
are deleted.

This patch retains the ids of all objects created to ensure
those exact objects are used and deleted as a part of the test.