ceph/rgw: Fine grained pool creation
We create the pools necessary for radosgw ourselves, this allows us to
control these settings in a more fine grained manner as different pools
can have different pg counts & replica counts. Also fixes #556, as the
new pools modules allows for pg_num to be increased which is a desirable
functionality, the cause of the issue being differing pg/pgp nums for
rgw pools.

Fixes: #556, #487
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@gmail.com>