Add the concept of an override_repo
This allows users to specify an external location where an
apt repo can be uploaded (like through swift, or on an apache server).

it also contains a script that can be used to generate
this repo.

This feature is intended to be used as a part of
a two part job, one to build a repo and set it as
an artifact, and one to use the artifact.

NOTE: this patch uses a module called archive which depends on
some packages, since it is setting up an apt::source, it
cannot install those packages first (b/c apt::source comes
before all packages). For this reason, those deps are installed
from userdata which might be less than ideal.