Add collectd to sensitive services
the collectd packages do not seem to install
cleanly. collectd depends on collectd-core,
which for some reason is trying to start the
collectd service before the collect package
writes a default config file.

This leads to the following error during package
install:

    Setting up collectd-core (5.4.0-3ubuntu2) ...
     * Starting statistics collection and monitoring daemon collectd
       ...fail!

Adds it to the list of sensitive services to work around this
failure.