refactor of consul collectd integration
This commit refactors the collectd/consul integration.

The previous attempt at the integration suffered from
multiple race conditions.

* multiple scripts were moving the same file, resulting
in failures for the case where they tried to move it at
the same time.
* each script processed the entire script, but ignored
anything, but it's own event types, this means that an
event type might miss all of it's own events if someone else
was always consuming them.

This refactor does the following:

- calling script is moved to cron and only called once
- check_threshold script is refactored to collect only
last event for each type
- checks are moved to ttl checks
- check_threshold script uses consul api to update checks