Use interface and ip address based on the subnet
Currently we assume that the ip address we need to configure in the cloud is
always setup on same interface on a specific set of servers/cloud provider,
which can be incorrect because of number of reasons
1. what if we had to use different nic port because of one bad port
2. compute node have the ip of physical interface will be moved to vhost0 in
which case you need special consideration - like custom facts

Rather the subnet is more consistent, this patch adds code to create custom
facts for ipaddress and interface based on subnet like interface_10_0_0_0_8 with
value eth0 in case eth0 have subnet configured 10.0.0.0/8.

This will allow us to use the fact ipaddress_{subnet} and interface_{subnet}
whereever required, as far as correct subnet provided, this will provide
consistant network config across the servers.

This should fix the issue #643.