Need more logic for swap_macs to work
swap_macs doesnt work with current code as the condition is happening in compile
time when the bridge interface would not be created, and thus it is breaking the
logic (because even though bridge interface would not be created, the interfaces
file will be created with appropriate config, which break the logic here).

So only create the interfaces file if bridge interface is already exists.