Swap the macs btn bridge and physical interface for gate
In case of gate or virtual environments, underlying contrail will send the
traffic to the mac of physical interface even if it dont have any IP, so the mac
addresses should be swapped in between physical interface and bridge interface
for gate/at.

Also removed unwanted code as they are handled in classess in module.