Moved rabbitmq call from rjil::ironic
rjil::ironic was calling the class ::rabbitmq, which is not desireable, so using
rjil::rabbitmq so all tests and customizations are included. Also moved it to
site.pp from rjil::ironic.

This is required to create rabbitmq admin user and set permissions.

TODO: only ironic and neutron is using rabbitmq right now, once this patch working would
need to work on using zmq for that.