Stage the image and convert if required before creating the image
This patch is to add facility to stage the image on local location and
optionaly convert it to raw format before uploading to glance.

This patch will solve two problems:

1. after we support only boot from volume, we need raw images, so this will
convert the images if asked to do so.
2. it use puppet-staging to stage the image which is internally using
curl/wget which respect global proxies, so it eleminate the need of floating IPs
to ocdb1.