Create common code for network interface configuration
Network interface configuration code need to be reused. I checked if we could
use the module puppet-network, but it dont have the configuration which we need.
Another reason is that we would need to refresh the network before writing the
configuration, which is not handled in the existing code in the module.

Added spec tests also