Changes related to baremetal uc and prestaging environment
Bunch of hiera changes to make this code with for baremetal system for undercloud controller. Also set of changes for prestaging to work.

NOTE: it need a custom fact for ipaddress_br_ctlplane_or_em4 required to get this worked, until the patch to implement subnetbased facts merged. It is not pushing right now assuming that the other patch will be merged soon.