Make flavor management optional
Sometimes you may not need to manage nova flavors, so making it optional. This
is useful in case of undercloud controller where the flavors would have
extra_specs which are not supported by nova_flavor at this moment (we currently
use openstack client in nova_flavor which doesnt support setting extra_specs).