Add encryption for consul
This patch adds encryption for consul
gossip by default. It uses the first
16 bytes of the consul discovery token
to ensure a reasonable configuration that
will be unique for each cluster.

It is also possible to override that
key using hiera.

This patch should solve a few things:

1. allows encryption of gossip traffic
2. ensures that only hosts that know
the shared secret can use gossip protocol

2 should also solve the accidental cluster
joining as a side effect

This feature is not enabled by default and
only enabled in the at env because it is
not clear how it can be upgraded.