Set shell as bash for files/tests/nagios_killall
This this script uses bash specific syntax, the
shell should be set as bash explicitly
to ensure that even if it is invoked as a
regular executable that it will run as expected.

This commit also adds set -e to ensure that the script
does not proceed if any of it fails.