Changes to conform subnet based interface selection
The patch https://github.com/JioCloud/puppet-rjil/pull/646 implemented facts
which allow to select the network interfaces/ipaddress based on subnet assigned.
This patch added the hiera changes for undercloud which conforms this.

1. changed the subnet for undercloud to make it same as overcloud - this will
eleminate the need of changing hiera data for uc and thus no need to have an
environment based role override.
2. Removed ipaddress_br_* facts references